Convert double to int in drawEnemy

+++ Packages/view/View.java
@@ -27,6 +27,6 @@ public class View {
         int y = 500;
         Image image = enemy.getImage();
         gc.drawImage(image, 0, 0, image.getWidth(), image.getHeight(),
-            image.getWidth() + enemy.getPosition() , y, -image.getWidth(), image.getHeight());
+            image.getWidth() + (int) enemy.getPosition() , y, -image.getWidth(), image.getHeight());
     }
 }
